refactor(web-todomvc-experimental): migrate to new multi-store API
