+++ examples/web-todomvc-experimental/package.json
@@ -13,6 +13,7 @@
     "@overengineering/fps-meter": "0.2.4",
     "react": "19.1.0",
     "react-dom": "19.1.0",
+    "react-error-boundary": "6.0.0",
     "todomvc-app-css": "2.4.3"
   },
   "devDependencies": {
