+++ examples/web-todomvc-experimental/src/Root.tsx
@@ -1,16 +1,12 @@
-import { makePersistedAdapter } from '@livestore/adapter-web'
-import LiveStoreSharedWorker from '@livestore/adapter-web/shared-worker?sharedworker'
-import { LiveStoreProvider } from '@livestore/react'
+import { StoreRegistry, StoreRegistryProvider } from '@livestore/react'
 import { FPSMeter } from '@overengineering/fps-meter'
 import type React from 'react'
-import { unstable_batchedUpdates as batchUpdates } from 'react-dom'
-
+import { Suspense, useState } from 'react'
+import { ErrorBoundary } from 'react-error-boundary'
 import { Footer } from './components/Footer.tsx'
 import { Header } from './components/Header.tsx'
 import { MainSection } from './components/MainSection.tsx'
 import { VersionBadge } from './components/VersionBadge.tsx'
-import { SyncPayload, schema } from './livestore/schema.ts'
-import LiveStoreWorker from './livestore.worker.ts?worker'
 
 const AppBody: React.FC = () => (
   <section className="todoapp">
@@ -20,27 +16,20 @@ const AppBody: React.FC = () => (
   </section>
 )
 
-const adapter = makePersistedAdapter({
-  storage: { type: 'opfs' },
-  worker: LiveStoreWorker,
-  sharedWorker: LiveStoreSharedWorker,
-})
+export const App: React.FC = () => {
+  const [registry] = useState(() => new StoreRegistry())
 
-const syncPayload = { authToken: 'insecure-token-change-me' }
-
-export const App: React.FC = () => (
-  <LiveStoreProvider
-    schema={schema}
-    renderLoading={(_) => <div>Loading LiveStore ({_.stage})...</div>}
-    adapter={adapter}
-    batchUpdates={batchUpdates}
-    syncPayloadSchema={SyncPayload}
-    syncPayload={syncPayload}
-  >
-    <div style={{ top: 0, right: 0, position: 'absolute', background: '#333' }}>
-      <FPSMeter height={40} />
-    </div>
-    <AppBody />
-    <VersionBadge />
-  </LiveStoreProvider>
-)
+  return (
+    <ErrorBoundary fallback={<div>Something went wrong</div>}>
+      <Suspense fallback={<div>Loading app...</div>}>
+        <StoreRegistryProvider storeRegistry={registry}>
+          <div style={{ top: 0, right: 0, position: 'absolute', background: '#333' }}>
+            <FPSMeter height={40} />
+          </div>
+          <AppBody />
+          <VersionBadge />
+        </StoreRegistryProvider>
+      </Suspense>
+    </ErrorBoundary>
+  )
+}
