+++ examples/web-todomvc-experimental/src/components/MainSection.tsx
@@ -1,11 +1,10 @@
 import type { Store } from '@livestore/livestore'
 import { queryDb } from '@livestore/livestore'
-import { useStore } from '@livestore/react'
 import { LiveList } from '@livestore/react'
 import React from 'react'
-
 import { uiState$ } from '../livestore/queries.ts'
 import { events, tables } from '../livestore/schema.ts'
+import { useAppStore } from '../livestore/store.ts'
 
 type Todo = typeof tables.todos.Type
 
@@ -21,7 +20,7 @@ const visibleTodos$ = queryDb(
 )
 
 export const MainSection: React.FC = () => {
-  const { store } = useStore()
+  const store = useAppStore()
 
   // We record an event that specifies marking complete or incomplete,
   // The reason is that this better captures the user's intention
