+++ examples/web-todomvc-experimental/src/livestore/store.ts
@@ -1,0 +1,24 @@
+import { makePersistedAdapter } from '@livestore/adapter-web'
+import LiveStoreSharedWorker from '@livestore/adapter-web/shared-worker?sharedworker'
+import { useStore } from '@livestore/react'
+import { unstable_batchedUpdates as batchUpdates } from 'react-dom'
+import LiveStoreWorker from '../livestore.worker.ts?worker'
+import { SyncPayload, schema } from './schema.ts'
+
+const adapter = makePersistedAdapter({
+  storage: { type: 'opfs' },
+  worker: LiveStoreWorker,
+  sharedWorker: LiveStoreSharedWorker,
+})
+
+const syncPayload = { authToken: 'insecure-token-change-me' }
+
+export const useAppStore = () =>
+  useStore({
+    storeId: 'app-root',
+    schema,
+    adapter,
+    batchUpdates,
+    syncPayloadSchema: SyncPayload,
+    syncPayload,
+  })
